+++ src/lib.rs
@@ -119,6 +119,7 @@ mod resource;
 mod param;
 mod payload;
 mod pipeline;
+mod with;
 mod extractor;
 
 pub mod client;
@@ -144,6 +145,7 @@ pub use route::Route;
 pub use resource::Resource;
 pub use context::HttpContext;
 pub use server::HttpServer;
+pub use extractor::{Path, Query, HttpRequestExtractor};
 
 // re-exports
 pub use http::{Method, StatusCode, Version};
@@ -184,10 +186,10 @@ pub mod dev {
     pub use context::Drain;
     pub use info::ConnectionInfo;
     pub use handler::Handler;
+    pub use with::With;
     pub use json::JsonBody;
     pub use router::{Router, Pattern};
     pub use param::{FromParam, Params};
     pub use httpmessage::{UrlEncoded, MessageBody};
     pub use httpresponse::HttpResponseBuilder;
-    pub use extractor::{Path, Query, HttpRequestExtractor};
 }